Extract out functions into own files.
