+++ assets/js/util/date-range/index.js
@@ -16,67 +16,6 @@
  * limitations under the License.
  */
 
-/**
- * WordPress dependencies
- */
-import { _n, sprintf } from '@wordpress/i18n';
-
-/**
- * Gets the current dateRange day count.
- *
- * @since 1.19.0
- * @since 1.26.0 `dateRange` is now a required argument.
- *
- * @param {string} dateRange The date range slug.
- * @return {number} The number of days in the range.
- */
-export function getCurrentDateRangeDayCount( dateRange ) {
-	const daysMatch = dateRange.match( /last-(\d+)-days/ );
-
-	if ( daysMatch && daysMatch[ 1 ] ) {
-		return parseInt( daysMatch[ 1 ], 10 );
-	}
-
-	throw new Error( 'Unrecognized date range slug.' );
-}
-
-/**
- * Gets the hash of available date ranges.
- *
- * @since 1.12.0
- *
- * @return {Object} The object hash where every key is a date range slug, and the value is an object with the date range slug and its translation.
- */
-export function getAvailableDateRanges() {
-	const label = ( days ) => sprintf(
-		/* translators: %s: number of days */
-		_n( 'Last %s day', 'Last %s days', days, 'google-site-kit', ),
-		days,
-	);
-
-	return {
-		'last-7-days': {
-			slug: 'last-7-days',
-			label: label( 7 ),
-			days: 7,
-		},
-		'last-14-days': {
-			slug: 'last-14-days',
-			label: label( 14 ),
-			days: 14,
-		},
-		'last-28-days': {
-			slug: 'last-28-days',
-			label: label( 28 ),
-			days: 28,
-		},
-		'last-90-days': {
-			slug: 'last-90-days',
-			label: label( 90 ),
-			days: 90,
-		},
-	};
-}
 export {
 	INVALID_DATE_INSTANCE_ERROR,
 	INVALID_DATE_RANGE_ERROR,
@@ -88,3 +27,5 @@ export { getPreviousWeekDate } from './get-previous-week-date';
 export { isValidDateInstance } from './is-valid-date-instance';
 export { isValidDateRange } from './is-valid-date-range';
 export { isValidDateString } from './is-valid-date-string';
+export { getCurrentDateRangeDayCount } from './get-current-date-range-day-count';
+export { getAvailableDateRanges } from './get-available-date-ranges';
